% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdEthio-package.R
\docType{package}
\name{cdEthio-package}
\alias{cdEthio}
\alias{cdEthio-package}
\title{cdEthio: Convert Date from the Ethiopian Calendar to the Gregorian}
\description{
The function in this package convert day, month, and year from the Ethiopian calendar to the Gregorian. It allows researchers to easily convert date in a large data set.
}
\author{
\strong{Maintainer}: Dawit Jember \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
