% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cddate.R
\name{davedate}
\alias{davedate}
\title{Convert date from the Ethiopian calendar to the Gregorian}
\usage{
davedate(a, b, c, id)
}
\arguments{
\item{a}{a numeric value of day DD}

\item{b}{a numeric value of month MM}

\item{c}{a numeric value of year YYYY}

\item{id}{a numeric value of individual id}
}
\value{
a date value
}
\description{
Convert date from the Ethiopian calendar to the Gregorian
}
\examples{
#The package has sample_date data
data(package="cdEthio")
#convert dates using the sample_date data
dateconverted<-davedate(a=sample_date$day, b=sample_date$month,
c=sampple_date$year, id=sample_date$id)
#check by entering date, month, year for a single individual
davedate(a=5, b=4, c=2013, id=1)
}
